+++ resources/views/livewire/events/event-overview.blade.php
@@ -2,6 +2,8 @@
     $hasEditPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\Event::EVENT_EDIT_PERMISSION) ?? false;
     $hasAttendanceShowPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\Attendance::ATTENDANCE_SHOW_PERMISSION) ?? false;
     $hasAttendanceEditPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\Attendance::ATTENDANCE_EDIT_PERMISSION) ?? false;
+    /** @var \App\Models\Filter\EventFilter $eventFilter */
+    $eventFilter = $this->getEventFilter();
 @endphp
 
 <div>
@@ -24,18 +26,18 @@
                 {{__("Calendar")}}
             </x-button-link>
         </div>
-            @if($hasEditPermission)
-                <div class="flex flex-row flex-wrap gap-2 justify-center">
-                    <x-button-link href="{{route('event.type.index')}}" class="btn-secondary"
-                                   title="Show event type list">
-                        {{__("Event Type List")}}
-                    </x-button-link>
-                    <x-button-link href="{{route('event.create')}}" class="btn-success"
-                                   title="Create new event">
-                        {{__("Create new event")}}
-                    </x-button-link>
-                </div>
-            @endif
+        @if($hasEditPermission)
+            <div class="flex flex-row flex-wrap gap-2 justify-center">
+                <x-button-link href="{{route('event.type.index')}}" class="btn-secondary"
+                               title="Show event type list">
+                    {{__("Event Type List")}}
+                </x-button-link>
+                <x-button-link href="{{route('event.create')}}" class="btn-success"
+                               title="Create new event">
+                    {{__("Create new event")}}
+                </x-button-link>
+            </div>
+        @endif
     </div>
 
     <x-livewire.loading/>
@@ -142,9 +144,11 @@
         {!! $eventList->links('vendor.livewire.paginator') !!}
     </div>
 
-    @if($hasEditPermission)
+    <div class="flex flex-col sm:flex-row gap-3 my-3 bg-white items-center shadow-sm sm:rounded-lg p-6 justify-between">
 
-        <div x-data="{clickCnt: 0, onClick(btn) {
+        @if($hasEditPermission)
+
+            <div x-data="{clickCnt: 0, onClick(btn) {
                 if(this.clickCnt > 0) {
                     btn.disabled = true;
                     $wire.disableLastYearEvents();
@@ -153,13 +157,49 @@
                     btn.innerHTML = 'Are you sure?';
                 }
             }}"
-             class="flex flex-col sm:flex-row gap-3 my-3 bg-white overflow-hidden items-center shadow-sm sm:rounded-lg p-6">
-            <x-default-button
-                x-on:click="onClick($el)" title="Disable all events older than this year."
-                class="btn-danger">{{ __('Disable last years events') }}</x-default-button>
-            @if(session()->has("eventDisableMessage"))
-                <span class="text-gray-700">{{session()->pull("eventDisableMessage")}}</span>
-            @endif
-        </div>
-    @endif
+                 >
+                <x-default-button
+                    x-on:click="onClick($el)" title="Disable all events older than this year."
+                    class="btn-danger">{{ __('Disable last years events') }}</x-default-button>
+                @if(session()->has("eventDisableMessage"))
+                    <span class="text-gray-700">{{session()->pull("eventDisableMessage")}}</span>
+                @endif
+            </div>
+        @endif
+
+
+        @auth
+            <div x-data="{
+                        open:false,
+                    }" class="relative inline-block text-left" @click.outside="open = false">
+                <div>
+                    <button type="button"
+                            class="inline-flex w-full justify-center items-center gap-x-1.5 rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50"
+                            @click.stop="open = !open">
+                        <i class="fa-solid fa-file-export"></i> Export
+                        <i class="fa-solid fa-chevron-down text-gray-400 transition"
+                           :class="open ? 'rotate-180' : ''"></i>
+                    </button>
+                </div>
+
+                <div x-cloak x-show="open"
+                     class="absolute left-full transform -translate-x-1/2 -translate-y-[9rem] z-10 mt-2 w-56 origin-top-right rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+                    <div class="py-1">
+                        <div class="px-4 py-1">
+                            <x-button-link class="w-full"
+                                           href="{{route('event.list.csv', $eventFilter->toParameterArray())}}"
+                                           @click="open=false"
+                                           title="Download event list as CSV file">{{ __('CSV List') }}</x-button-link>
+                        </div>
+                        <div class="px-4 py-1">
+                            <x-button-link class="w-full"
+                                           href="{{route('event.list.excel', $eventFilter->toParameterArray())}}"
+                                           @click="open=false"
+                                           title="Download event list as Excel file">{{ __('Excel File') }}</x-button-link>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        @endauth
+    </div>
 </div>